backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-dotenv==1.1.1
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
openpyxl==3.1.2
openai==1.51.2
